+++ lib/src/query/raw_query_builder.dart
@@ -9,6 +9,13 @@ import '../../delta_trace_db.dart';
 class RawQueryBuilder extends QueryBuilder {
   List<Map<String, dynamic>>? rawAddData;
 
+  /// (en) Adds an item to the specified collection.
+  /// If the specified collection does not already exist,
+  /// it will be created automatically.
+  ///
+  /// (ja) 指定されたコレクションに要素を追加します。
+  /// 指定されたコレクションがまだ存在しない場合はコレクションが自動で作成されます。
+  ///
   /// * [target] : The collection name in DB.
   /// * [rawAddData] : Data specified when performing an add operation.
   /// * [returnData] : If true, return the changed objs.
@@ -39,6 +46,13 @@ class RawQueryBuilder extends QueryBuilder {
   }) : this.rawAddData = rawAddData,
        super.add(addData: []);
 
+  /// (en) Overwrites the parameters of all objects in the specified collection
+  /// that match the conditions.
+  /// Parameters not specified for overwriting remain unchanged.
+  ///
+  /// (ja) 指定されたコレクションのうち、条件にマッチする全てのオブジェクトのパラメータを
+  /// 上書きします。上書き対象に指定していないパラメータは変化しません。
+  ///
   /// * [target] : The collection name in DB.
   /// * [queryNode] : This is the node object used for the search.
   /// You can build queries by combining the various nodes defined in
@@ -74,6 +88,13 @@ class RawQueryBuilder extends QueryBuilder {
     super.cause,
   }) : super.update();
 
+  /// (en) Overwrites the parameters of one object in the specified collection
+  /// that matches the conditions. Parameters not specified for overwriting
+  /// remain unchanged.
+  ///
+  /// (ja) 指定されたコレクションのうち、条件にマッチする１つのオブジェクトのパラメータを
+  /// 上書きします。上書き対象に指定していないパラメータは変化しません。
+  ///
   /// * [target] : The collection name in DB.
   /// * [queryNode] : This is the node object used for the search.
   /// You can build queries by combining the various nodes defined in
@@ -104,6 +125,11 @@ class RawQueryBuilder extends QueryBuilder {
     super.cause,
   }) : super.updateOne();
 
+  /// (en) Deletes all objects in the specified collection that match
+  /// the specified criteria.
+  ///
+  /// (ja) 指定されたコレクションのうち、条件にマッチするオブジェクトを全て削除します。
+  ///
   /// * [target] : The collection name in DB.
   /// * [queryNode] : This is the node object used for the search.
   /// You can build queries by combining the various nodes defined in
@@ -130,6 +156,11 @@ class RawQueryBuilder extends QueryBuilder {
     super.cause,
   }) : super.delete();
 
+  /// (en) Deletes only one object that matches the specified criteria from
+  /// the specified collection.
+  ///
+  /// (ja) 指定されたコレクションのうち、条件にマッチするオブジェクトを１件だけ削除します。
+  ///
   /// * [target] : The collection name in DB.
   /// * [queryNode] : This is the node object used for the search.
   /// You can build queries by combining the various nodes defined in
@@ -153,6 +184,11 @@ class RawQueryBuilder extends QueryBuilder {
     super.cause,
   }) : super.deleteOne();
 
+  /// (en) Gets objects from the specified collection that match
+  /// the specified criteria.
+  ///
+  /// (ja) 指定されたコレクションから、条件にマッチするオブジェクトを取得します。
+  ///
   /// * [target] : The collection name in DB.
   /// * [queryNode] : This is the node object used for the search.
   /// You can build queries by combining the various nodes defined in
@@ -199,6 +235,10 @@ class RawQueryBuilder extends QueryBuilder {
     super.cause,
   }) : super.search();
 
+  /// (en) Gets all items in the specified collection.
+  ///
+  /// (ja) 指定されたコレクションの全てのアイテムを取得します。
+  ///
   /// * [target] : The collection name in DB.
   /// * [sortObj] : An object for sorting the return values.
   /// SingleSort or MultiSort can be used.
@@ -211,6 +251,16 @@ class RawQueryBuilder extends QueryBuilder {
   RawQueryBuilder.getAll({required super.target, super.sortObj, super.cause})
     : super.getAll();
 
+  /// (en) Formats the contents of the specified collection to match the
+  /// specified template.
+  /// Fields not present in the template will be removed,
+  /// and fields that are only present in the template will be added
+  /// with the template's value as their initial value.
+  ///
+  /// (ja) 指定されたコレクションの内容を、指定したテンプレートに一致するように整形します。
+  /// テンプレートに存在しないフィールドは削除され、テンプレートにのみ存在するフィールドは、
+  /// テンプレートの値を初期値として追加されます。
+  ///
   /// * [target] : The collection name in DB.
   /// * [template] : Specify this when changing the structure of the DB class.
   /// Fields that do not exist in the existing structure but exist in the
@@ -233,6 +283,10 @@ class RawQueryBuilder extends QueryBuilder {
     super.cause,
   }) : super.conformToTemplate(template: template);
 
+  /// (en) Renames a specific field in the specified collection.
+  ///
+  /// (ja) 指定されたコレクションの特定のフィールドの名前を変更します。
+  ///
   /// * [target] : The collection name in DB.
   /// * [renameBefore] : The old variable name when querying for a rename.
   /// * [renameAfter] : The new name of the variable when querying for a rename.
@@ -256,6 +310,10 @@ class RawQueryBuilder extends QueryBuilder {
     super.cause,
   }) : super.renameField();
 
+  /// (en) Gets the number of elements in the specified collection.
+  ///
+  /// (ja) 指定されたコレクションの要素数を取得します。
+  ///
   /// * [target] : The collection name in DB.
   /// * [cause] : You can add further parameters such as why this query was
   /// made and who made it.
@@ -265,6 +323,10 @@ class RawQueryBuilder extends QueryBuilder {
   /// the DB history is recorded.
   RawQueryBuilder.count({required super.target, super.cause}) : super.count();
 
+  /// (en) Clears the specified collection.
+  ///
+  /// (ja) 指定されたコレクションをclearします。
+  ///
   /// * [target] : The collection name in DB.
   /// * [mustAffectAtLeastOne] : If true, the operation will be marked as
   /// failed if it affects 0 objects.
@@ -285,6 +347,10 @@ class RawQueryBuilder extends QueryBuilder {
     super.cause,
   }) : super.clear();
 
+  /// (en) Clears the specified collection and then add data.
+  ///
+  /// (ja) 指定されたコレクションをclearした後、dataをAddします。
+  ///
   /// * [target] : The collection name in DB.
   /// * [rawAddData] : Data specified when performing an add operation.
   /// * [returnData] : If true, return the changed objs.